move registration summaries
