+++ embc-app/ClientApp/src/app/app.module.ts
@@ -13,14 +13,14 @@ import { AppRoutingModule } from './app-routing.module';
 import { CoreModule } from './core/core.module';
 import { SelfRegistrationModule } from './self-registration';
 import { SharedModule } from './shared/shared.module';
-import { RegistrationSummaryComponent } from './registration-summary/registration-summary.component';
+import { RegistrationSummaryComponent } from './components/registration-summary/registration-summary.component';
 import { TaskNumberListComponent } from './provincial-admin/components/task-number-list/task-number-list.component';
 import { VolunteerListComponent } from './local-authority/components/volunteer-list/volunteer-list.component';
 import { OrganizationListComponent } from './provincial-admin/components/organization-list/organization-list.component';
 import { TaskNumberMakerComponent } from './provincial-admin/components/task-number-maker/task-number-maker.component';
 import { VolunteerMakerComponent } from './local-authority/components/volunteer-maker/volunteer-maker.component';
 import { OrganizationMakerComponent } from './provincial-admin/components/organization-maker/organization-maker.component';
-import { RegistrationSummaryFullComponent } from './registration-summary-full/registration-summary-full.component';
+import { RegistrationSummaryFullComponent } from './components/registration-summary-full/registration-summary-full.component';
 import { VolunteerOrganizationListComponent } from './provincial-admin/components/volunteer-organization-list/volunteer-organization-list.component';
 import { RegistrationMakerComponent } from './volunteer/components/registration-maker/registration-maker.component';
 import { SessionExpiredComponent } from './session-expired/session-expired.component';
