+++ embc-app/ClientApp/src/app/components/registration-summary-full/registration-summary-full.component.html
