+++ embc-app/ClientApp/src/app/components/registration-summary/registration-summary.component.scss
